# docs/api/v1/openapi.yaml
openapi: 3.1.0
info:
  title: FreshLens API
  version: 1.0.0
  description: |
    V1 contract for FreshLens (CS3203 · Group 21 · PID 5).

    - Base path for versioned resources: `/api/v1`
    - Auth: Supabase JWT Bearer on versioned routes except public tenant
      application submission; `GET /health` is also public
    - Tenant isolation: auth middleware derives trusted identity from the JWT, and
      the tenant DB transaction sets Postgres `app.tenant_id` on its own connection;
      never trust a client-supplied `tenant_id`
    - Async scan only: `POST /api/v1/scans` returns **202**; the two-tier CNN
      pipeline runs in Celery, never in the request handler

    See `docs/api/README.md` for human overview and out-of-scope notes.
servers:
  - url: http://localhost:8000
    description: Local Docker Compose / FastAPI (prototype)
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/scans:
    post:
      tags: [Scans]
      summary: Submit a produce scan (async)
      description: |
        Accepts one product photo plus vendor-confirmed quantity.
        Uploads image (e.g. R2), inserts a `pending` scan row, enqueues a Celery job,
        and returns **202** immediately. Clients poll `GET /api/v1/scans/{scan_id}`.
      operationId: createScan
      security:
        - bearerAuth: []
      x-freshlens-roles: [vendor, tenant_admin]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/CreateScanMultipart"
      responses:
        "202":
          description: Scan accepted; inference queued
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanAccepted"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "422":
          $ref: "#/components/responses/ValidationError"
    get:
      tags: [Scans]
      summary: List recent scans for the current tenant
      operationId: listScans
      security:
        - bearerAuth: []
      x-freshlens-roles: [vendor, tenant_admin]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Tenant-scoped scan page
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScanList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/scans/{scan_id}:
    get:
      tags: [Scans]
      summary: Get a scan by id
      description: |
        Returns the full scan resource. While status is `pending` or `processing`,
        classification fields may be null. For an accepted identity, a completed
        scan includes the Model 2 classification and freshness score. Rejected
        identities complete without a freshness grade.
      operationId: getScan
      security:
        - bearerAuth: []
      x-freshlens-roles: [vendor, tenant_admin]
      parameters:
        - $ref: "#/components/parameters/ScanId"
      responses:
        "200":
          description: Scan resource
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scan"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/alerts:
    get:
      tags: [Alerts]
      summary: List alerts for the current tenant
      description: |
        Returns tenant-scoped, actionable inventory alerts. Active alerts are
        returned by default; resolved history can be requested explicitly.
      operationId: listAlerts
      security:
        - bearerAuth: []
      x-freshlens-roles: [vendor, tenant_admin]
      parameters:
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - name: active_only
          in: query
          required: false
          schema: { type: boolean, default: true }
      responses:
        "200":
          description: Tenant-scoped alerts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AlertList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/alerts/{alert_id}/read:
    patch:
      tags: [Alerts]
      summary: Mark one tenant alert as read
      operationId: markAlertRead
      security:
        - bearerAuth: []
      x-freshlens-roles: [vendor, tenant_admin]
      parameters:
        - name: alert_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      responses:
        "200":
          description: Updated tenant alert
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Alert"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /api/v1/tenant-applications:
    post:
      tags: [Tenant applications]
      summary: Submit a tenant signup application
      operationId: submitTenantApplication
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantApplicationCreate"
      responses:
        "202":
          description: Application accepted for platform review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantApplicationSubmitted"
        "422":
          $ref: "#/components/responses/ValidationError"
  /api/v1/admin/tenant-applications:
    get:
      tags: [Tenant applications]
      summary: List tenant signup applications
      operationId: listTenantApplications
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/TenantApplicationStatus"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Application review queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantApplicationList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
  /api/v1/admin/tenant-applications/{application_id}/approve:
    post:
      tags: [Tenant applications]
      summary: Approve an application, create its tenant, and invite its owner
      operationId: approveTenantApplication
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantApplicationReview"
      responses:
        "200":
          description: Approved application linked to its tenant and owner
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantApplication"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Application is no longer pending or email already exists
  /api/v1/admin/tenant-applications/{application_id}/reject:
    post:
      tags: [Tenant applications]
      summary: Reject a pending tenant application
      operationId: rejectTenantApplication
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: application_id
          in: path
          required: true
          schema: { $ref: "#/components/schemas/Uuid" }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TenantApplicationReview"
      responses:
        "200":
          description: Rejected application
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TenantApplication"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Application is no longer pending
...
  /api/v1/admin/alerts:
    get:
      tags: [Admin]
      summary: List tenant alert signals
      description: Returns alert signals and product context without batch identifiers.
      operationId: listAdminAlerts
      security:
        - bearerAuth: []
      x-freshlens-roles: [platform_admin]
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 100
        - $ref: "#/components/parameters/Offset"
        - name: search
          in: query
          schema: { type: string, maxLength: 100, default: "" }
          description: Case-insensitive alert message, tenant, or product search.
        - name: alert_type
          in: query
          schema: { $ref: "#/components/schemas/AlertType" }
        - name: severity
          in: query
          schema: { $ref: "#/components/schemas/AlertSeverity" }
      responses:
        "200":
          description: Tenant alert signals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminAlertList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Supabase Auth access token (JWT).
  parameters:
    ScanId:
      name: scan_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/Uuid"
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      required: false
      schema:
        type: integer
        minimum: 0
        default: 0
  responses:
    Unauthorized:
      description: Missing or invalid JWT
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but wrong role or tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Resource not found in this tenant
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
    Uuid:
      type: string
      format: uuid
...
    TenantApplicationStatus:
      type: string
      enum: [pending, approved, rejected]
    TenantApplicationCreate:
      type: object
      required: [organization_name, applicant_name, applicant_email]
      additionalProperties: false
      properties:
        organization_name: { type: string, minLength: 2, maxLength: 120 }
        applicant_name: { type: string, minLength: 2, maxLength: 120 }
        applicant_email: { type: string, format: email, maxLength: 254 }
        phone:
          type: [string, "null"]
          minLength: 5
          maxLength: 40
    TenantApplicationSubmitted:
      type: object
      required: [id, status, message]
      additionalProperties: false
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        status: { const: pending }
        message: { type: string }
    TenantApplicationReview:
      type: object
      additionalProperties: false
      properties:
        note:
          type: [string, "null"]
          maxLength: 1000
    TenantApplication:
      type: object
      additionalProperties: false
      required: [id, organization_name, applicant_name, applicant_email, phone, status, review_note, reviewed_by, approved_tenant_id, approved_user_id, submitted_at, reviewed_at, updated_at]
      properties:
        id: { $ref: "#/components/schemas/Uuid" }
        organization_name: { type: string }
        applicant_name: { type: string }
        applicant_email: { type: string, format: email }
        phone: { type: [string, "null"] }
        status: { $ref: "#/components/schemas/TenantApplicationStatus" }
        review_note: { type: [string, "null"] }
        reviewed_by:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        approved_tenant_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        approved_user_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        submitted_at: { type: string, format: date-time }
        reviewed_at: { type: [string, "null"], format: date-time }
        updated_at: { type: string, format: date-time }
    TenantApplicationList:
      type: object
      required: [items, total, limit, offset]
      additionalProperties: false
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/TenantApplication" }
        total: { type: integer, minimum: 0 }
        limit: { type: integer, minimum: 1 }
        offset: { type: integer, minimum: 0 }
    Classification:
      type: string
      enum: [fresh, medium, spoiled]
      description: FL-2TC Tier-2 freshness label.
    ScanStatus:
      type: string
      enum: [pending, processing, completed, failed]
    AlertSeverity:
      type: string
      enum: [info, warning, critical]
    AlertType:
      type: string
      enum: [spoilage, low_stock, aging, other]
...
    Error:
      type: object
      required: [detail]
      additionalProperties: false
      properties:
        detail:
          description: FastAPI-compatible error payload (string or structured object).
          oneOf:
            - type: string
            - type: object
              additionalProperties: true
...
    CreateScanMultipart:
      type: object
      required: [image, quantity]
      additionalProperties: false
      properties:
        image:
          type: string
          format: binary
          description: Single produce photo (V1 = one product per photo).
        quantity:
          type: integer
          minimum: 1
          description: Vendor-confirmed quantity at scan time.
        product_id:
          $ref: "#/components/schemas/Uuid"
        batch_id:
          $ref: "#/components/schemas/Uuid"
    ScanAccepted:
      type: object
      required: [id, status, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        status:
          $ref: "#/components/schemas/ScanStatus"
          description: Typically `pending` at acceptance time.
        created_at:
          type: string
          format: date-time
    Scan:
      type: object
      required:
        - id
        - tenant_id
        - status
        - image_path
        - quantity
        - created_at
        - updated_at
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        tenant_id:
          $ref: "#/components/schemas/Uuid"
          description: Echo of tenant from JWT / RLS context (not client-authored).
        status:
          $ref: "#/components/schemas/ScanStatus"
        image_path:
          type: string
          description: Object storage key or URL path for the uploaded image.
        quantity:
          type: integer
          minimum: 1
        classification:
          oneOf:
            - $ref: "#/components/schemas/Classification"
            - type: "null"
        freshness_score:
          oneOf:
            - type: number
              format: float
              minimum: 0
              maximum: 1
            - type: "null"
          description: Model confidence or score; null until completed.
        model_version:
          oneOf:
            - type: string
            - type: "null"
          description: Combined identity and freshness pipeline version when both models run.
        identity_label:
          oneOf:
            - type: string
              enum: [Banana, Cucumber, Eggplant, Tomato]
            - type: "null"
          description: Tier-1 catalogue-compatible label; null for unknown or rejected input.
        identity_score:
          oneOf:
            - type: number
              format: float
              minimum: 0
              maximum: 1
            - type: "null"
          description: Tier-1 identity confidence; distinct from freshness_score.
        identity_model_version:
          oneOf:
            - type: string
            - type: "null"
        product_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        event_key:
          type: [string, "null"]
        product_name:
          type: [string, "null"]
        quantity_received:
          type: [integer, "null"]
          minimum: 0
        quantity_remaining:
          type: [integer, "null"]
          minimum: 0
        transition_at:
          type: [string, "null"]
          format: date-time
        read_at:
          type: [string, "null"]
          format: date-time
        resolved_at:
          type: [string, "null"]
          format: date-time
        batch_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ScanList:
      type: object
      required: [items, total, limit, offset]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Scan"
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
    Alert:
      type: object
      required: [id, type, message, severity, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        type:
          $ref: "#/components/schemas/AlertType"
        message:
          type: string
        severity:
          $ref: "#/components/schemas/AlertSeverity"
        created_at:
          type: string
          format: date-time
        batch_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        product_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
...
    AlertList:
      type: object
      required: [items, total, limit, offset]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Alert"
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
...
    AdminAlert:
      type: object
      required: [id, tenant_id, tenant_name, type, severity, message, created_at]
      additionalProperties: false
      properties:
        id:
          $ref: "#/components/schemas/Uuid"
        tenant_id:
          $ref: "#/components/schemas/Uuid"
        tenant_name:
          type: string
        type:
          $ref: "#/components/schemas/AlertType"
        severity:
          $ref: "#/components/schemas/AlertSeverity"
        message:
          type: string
        product_id:
          oneOf:
            - $ref: "#/components/schemas/Uuid"
            - type: "null"
        product_name:
          oneOf:
            - type: string
            - type: "null"
        created_at:
          type: string
          format: date-time
    AdminAlertList:
      type: object
      required: [items, total, limit, offset]
      additionalProperties: false
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/AdminAlert"
        total:
          type: integer
          minimum: 0
        limit:
          type: integer
          minimum: 1
        offset:
          type: integer
          minimum: 0
...